Added modules, fixed url of photos
